requirements.txt
discord.py[voice]==2.3.2
aiohttp==3.9.3
python-dotenv==1.0.1
apscheduler==3.10.4
audioop-lts==0.2.1; python_version>="3.13"
PyNaCl==1.7.1
